feat: use curl as fallback
Previously, curl would only be used if gh or glab were not found. But if
the executable was found it was used.

Now, if the the executable returns a non zero exit code, a curl request
will be send. This is specifically useful, if gh or glab are installed
but not setup correctly.