Insert unabbreviated commit hash
When rendering commit messages on forges (at least GitHub), the commit
hash is automatically truncated and hyperlinked. However, if the commit
hash is already truncated then the forge does not recognise it as such
and so does not hyperlink it.

This change inserts the full, unabbreviated commit hash in order to
leverage this forge behaviour. It does not change the completion list
which still displays the abbreviated commit hash.